Complement Demo
* Implement ViewModel
* Handle SMS send process an update UI State
* Notify user of SMS status
