+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.SMSenderLibrary"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".SMSenderActivity"
             android:exported="true"
             android:theme="@style/Theme.SMSenderLibrary">
             <intent-filter>
