+++ app/src/main/java/com/apisap/smsenderlibrary/ui/state/SMSenderUiState.kt
@@ -1,0 +1,3 @@
+package com.apisap.smsenderlibrary.ui.state
+
+data class SMSenderUiState(var isSendingSMS: Boolean = false)
